Upgraded the rans_static4k benchmark harness (-t).
It now does round-trip checking.  This slowed it up though (and
obviously all my other rans variants).  To solve this it now
preallocates and uses a different interface that avoids mallocing
data.  This is actually a more useful interface, so worth having.

Also added code to do memory warmup (idea taken from FSE) as this has
a significant speed up.  Whether or not it's "valid" benchmarking I
don't know, but it's directly comparable to FSE.

TODO: update the other rans variants so we're comparing like with
like.