Bug fix to rANS_static4_16i.c and also major speed up. The corruptions were due to (as advertised in the code) the integer division reciprocals only being valid on 31-bit numbers and not 32-bit.
The code also now uses an assembly version of RansDecRenorm.  If
you're using gcc it's possible that one of the non-assembly variable
may be faster still.  See the comments.

The default table size is now 2^9 instead of 2^10 for order-1.  This
is a noticable drop in accuracy (around 1% for the data I tried), but
considerably improves speed due to smaller table sizes fitting better
in cache.  Feel free to tweak TF_SHIFT_O1 yourself.

Also small tweaks to README benchmarks and Makefile.