An experimental improvement to 4k order-1 decoding.
We use a map to group the observed characters together rather than
spreading across the D[] space.  This introduces an extra indirection,
but on some system/compiler combinations is as much as 5% faster. (On
others it's perhaps 5% slower.)  Hit and miss, but for my default
setup it's a win.  Maybe we can figure out which is best
programmatically?