Fixed potential crashes with reading past input buffer.
The use of a branchless RansDecRenorm is all well and good, but it
reads the next byte of the data regardless of whether it is consumed.
This means we always read 1 beyond the end of the input buffer.  While
we could add a requirement that all input must be 1 byte larger, it
only works out as a ~2% hit (even less in order 1) to just do the
check.

Also culled the slower rans_uncompress_O1 variants.