Fixed a bug in the order-1 encoder fir r32x16b_avx2.
I'd baked the -1 (for > vs >=) into xmax, but forgot to update the
order-1 codec.

Also tweaked the Makefile so the default block size is not a power of
2, helping the order-1 decode speeds considerably.  Order-1 requires
the input buffer to be split into N equal sized chunks, which causes
cache line collisions if the division gives the same low order bits
regularly.

Finally TF_SHIFT and TF_SHIFT_O1 are now settable on the command line
with -DTF_SHIFT_O1=10 (for example).  Set the default O1 size to 10
for r32x16b_avx2 as if we're using order-1 we're already saying we
want a bit better ratio instead of speed, and 9 is pushing it too low
on (eg) bwt data.